добавил catalog экран, перевёл проект с версии котлина 1.9.0 на 2.0.0

+++ app/src/main/java/com/example/pizza_shift_2025/data/model/SizeModel.kt
@@ -2,5 +2,5 @@ package com.example.pizza_shift_2025.data.model
 
 data class SizeModel(
     val name: SizeTypeModel,
-    val price: Int
+    val price: Double
 )
